Refactor MonacoEditor in new file
